refactor: move provider types to @/shared/api/providers
