+++ keep-ui/features/workflows/builder/lib/utils.tsx
@@ -1,4 +1,4 @@
-import { Provider } from "@/app/(keep)/providers/providers";
+import { Provider } from "@/shared/api/providers";
 import {
   ToolboxConfiguration,
   V2StepConditionThreshold,
