+++ tests/conftest.py
@@ -543,7 +543,7 @@ def browser():
         browser = p.chromium.launch(headless=headless)
         context = browser.new_context()
         page = context.new_page()
-        page.set_default_timeout(5000)
+        page.set_default_timeout(50000)
         yield page
         context.close()
         browser.close()
